# lib/resupply-api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title — orval uses it to name generated output files.
  title: Api
  version: 0.1.0
  description: |
    CPAP Resupply Automation admin API. Powers the admin dashboard
    (queue, episode detail, conversation viewer, fulfillment). Patient
    PHI flows through this API and lives in the resupply.* schema.
servers:
  - url: /resupply-api
    description: Base API path
...
paths:
...
  /patients/{id}/notes:
    get:
      operationId: listPatientNotes
      tags: [patients]
      summary: Recent admin case-notes for a patient
      description: |
        Returns up to the 50 most-recent notes for the patient,
        decrypted server-side. Notes are append-only by design —
        there is no PATCH or DELETE endpoint, since rewriting
        history defeats their purpose as an admin-handoff log.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notes (newest first, max 50).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientNotesPage"
        "404":
          description: No patient with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
    post:
      operationId: createPatientNote
      tags: [patients]
      summary: Append a case-note to the patient record
      description: |
        Appends a free-text note to the patient. Body is encrypted
        at the SQL site via pgcrypto. Audit row records the body
        length only, never the contents.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientNoteRequest"
      responses:
        "201":
          description: Note created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientNote"
        "400":
          description: Body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleValidationError"
        "404":
          description: No patient with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /patients/{id}/prescriptions:
    post:
      operationId: createPrescription
      tags: [patients]
      summary: Record a new prescription for a patient
      description: |
        Creates a new prescription row, encrypting the optional
        `details` JSON (prescriber name/NPI, diagnosis, free-text
        notes) at the SQL site. The prescription starts in
        `active` status. Once recorded, clinical fields are
        immutable — the only legal mutation is a status transition
        via `PATCH /prescriptions/{rxId}`.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrescriptionRequest"
      responses:
        "201":
          description: Prescription created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePrescriptionResponse"
        "400":
          description: Body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleValidationError"
        "404":
          description: No patient with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /prescriptions/{rxId}:
    patch:
      operationId: updatePrescriptionStatus
      tags: [patients]
      summary: Transition a prescription's lifecycle status
      description: |
        The only mutation allowed on a prescription after creation.
        Accepts `status: active | expired | revoked` and writes
        nothing else. Clinical fields are immutable to keep the
        provenance honest — to "edit" a prescription, create a new
        one and mark the prior one expired.
      parameters:
        - in: path
          name: rxId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePrescriptionStatusRequest"
      responses:
        "200":
          description: Status updated (or no-op if unchanged).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatePrescriptionStatusResponse"
        "400":
          description: Body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleValidationError"
        "404":
          description: No prescription with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
...
components:
  schemas:
...
    AuthError:
      type: object
      description: |
        Auth failure body. Carries a single human-readable `error`
        string safe for the dashboard to render verbatim. Never
        contains the underlying error message, the user id, or any
        connection details.
      properties:
        error:
          type: string
      required:
        - error
...
    ConsoleValidationError:
      type: object
      description: Returned when a list endpoint's query string fails zod validation.
      properties:
        error:
          type: string
          enum: [invalid_query]
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required: [path, message]
      required:
        - error
        - issues
    NotFoundError:
      type: object
      description: Returned when a {id} path parameter does not match any row.
      properties:
        error:
          type: string
          enum: [not_found]
      required:
        - error
...
    PatientNote:
      type: object
      description: |
        A single admin case-note. Body is decrypted server-side
        before serialization; the wire format is plaintext but
        the column is encrypted at rest.
      properties:
        id:
          type: string
          format: uuid
        body:
          type: string
        authorEmail:
          type: string
        authorUserId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - body
        - authorEmail
        - createdAt
    PatientNotesPage:
      type: object
      description: |
        The list response is wrapped (rather than a bare array) so we
        can add filters / pagination later without a breaking change.
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PatientNote"
        count:
          type: integer
          minimum: 0
          description: Length of items[] (max 50, newest first).
      required:
        - items
        - count
    CreatePatientNoteRequest:
      type: object
      description: |
        Body for POST /patients/{id}/notes. Free-text; the column is
        encrypted at the SQL site via pgcrypto so the body never lands
        in plaintext on disk or in the audit metadata.
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 4000
      required:
        - body
      additionalProperties: false
    CreatePrescriptionRequest:
      type: object
      description: |
        Body for POST /patients/{id}/prescriptions. The optional
        `details` fields (prescriberName, prescriberNpi, diagnosis,
        notes) are stitched into the encrypted `details` JSON column;
        encrypt happens server-side.
      properties:
        itemSku:
          type: string
          minLength: 1
          maxLength: 64
        cadenceDays:
          type: integer
          minimum: 1
          maximum: 365
        validFrom:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        validUntil:
          type: string
          nullable: true
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        prescriberName:
          type: string
          nullable: true
          maxLength: 160
        prescriberNpi:
          type: string
          nullable: true
          maxLength: 20
        diagnosis:
          type: string
          nullable: true
          maxLength: 2000
        notes:
          type: string
          nullable: true
          maxLength: 2000
      required:
        - itemSku
        - cadenceDays
        - validFrom
      additionalProperties: false
    CreatePrescriptionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
      required:
        - id
    UpdatePrescriptionStatusRequest:
      type: object
      description: |
        Body for PATCH /prescriptions/{rxId}. Only the lifecycle
        status is mutable — clinical fields are immutable on
        purpose to keep provenance honest.
      properties:
        status:
          type: string
          enum: [active, expired, revoked]
      required:
        - status
      additionalProperties: false
    UpdatePrescriptionStatusResponse:
      type: object
      description: |
        Echoes the prescription id and the resulting status.
        `changed` is false when the requested status equals the
        current one (no-op idempotency for double-clicks).
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, expired, revoked]
        changed:
          type: boolean
      required:
        - id
        - status
        - changed
...